lib-storage: Add mail_location=..:ITERINDEX
This changes mailbox list iteration to work using INDEX directory instead of
the normal mail directory. This can be helpful when the indexes are stored
on a faster storage.

TODO: This works only with LAYOUT=fs for now. INBOX file handling might also
not be exactly correct in all situations, but that typically affects only
mbox format.