log: Fix busy loop when a long line is logged
If the line exceeded PIPE_MAX (4 kB usually), the process got into a busy
loop. It would still break out of it every 100ms and maybe handle some
other services' logging, but the service that logged the long line would
be stuck. This resulted in all those service processes also hanging because
they were waiting to be able to write to log.